eStargz: refactor codes by flattening and deduplicating
Initially, we've wrapped "wrapped" stargz lib (github.com/google/crfs/stargz)
for acheving eStargz spec (e.g. verification). This extension was done by
decoding and modifying (encoded) stargz blob produced by stargz lib.

Currently we maintains our own fork of stargz lib so we can extend the
functionality directly, without "wrapping" (i.e. decoding+modifying). This
commit refactors estargz lib to directly extend stargz lib with eliminating
wrapper codes and flatten the directory structure.
