+++ estargz/build.go
@@ -14,29 +14,32 @@
    limitations under the License.
 */
 
+/*
+   Copyright 2019 The Go Authors. All rights reserved.
+   Use of this source code is governed by a BSD-style
+   license that can be found in the LICENSE file.
+*/
+
 package estargz
 
 import (
 	"archive/tar"
 	"bytes"
+	"compress/gzip"
+	"encoding/json"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"os"
 	"path"
 	"runtime"
 	"strings"
 	"sync"
 
-	"github.com/containerd/stargz-snapshotter/estargz/stargz"
-	"github.com/containerd/stargz-snapshotter/estargz/tempfiles"
-	"github.com/containerd/stargz-snapshotter/estargz/positionwatcher"
+	"github.com/hashicorp/go-multierror"
 	digest "github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
-)
-
-const (
-	landmarkContents = 0xf
 )
 
 type options struct {
@@ -45,6 +48,7 @@ type options struct {
 
 type Option func(o *options)
 
+// WithChunkSize option specifies the chunk size of eStargz blob to build.
 func WithChunkSize(chunkSize int) Option {
 	return func(o *options) {
 		o.chunkSize = chunkSize
@@ -61,7 +65,7 @@ func Build(tarBlob *io.SectionReader, prioritized []string, opt ...Option) (_ io
 	for _, o := range opt {
 		o(&opts)
 	}
-	layerFiles := tempfiles.NewTempFiles()
+	layerFiles := newTempFiles()
 	defer func() {
 		if rErr != nil {
 			if err := layerFiles.CleanupAll(); err != nil {
@@ -69,54 +73,32 @@ func Build(tarBlob *io.SectionReader, prioritized []string, opt ...Option) (_ io
 			}
 		}
 	}()
-	entries, err := sort(tarBlob, prioritized)
+	entries, err := sortEntries(tarBlob, prioritized)
 	if err != nil {
 		return nil, "", err
 	}
 	tarParts := divideEntries(entries, runtime.GOMAXPROCS(0))
-	sgzParts := make([]*io.SectionReader, len(tarParts))
-	var sgzPartsMu sync.Mutex
+	writers := make([]*Writer, len(tarParts))
+	payloads := make([]*os.File, len(tarParts))
+	var mu sync.Mutex
 	var eg errgroup.Group
 	for i, parts := range tarParts {
 		i, parts := i, parts
 		// builds verifiable stargz sub-blobs
 		eg.Go(func() error {
-			tmp := tempfiles.NewTempFiles()
-			defer tmp.CleanupAll()
-			tmpFile, err := tmp.TempFile("", "tmpdata")
-			if err != nil {
-				return err
-			}
-			sw := stargz.NewWriter(tmpFile)
-			sw.ChunkSize = opts.chunkSize
-			if err := sw.AppendTar(readerFromEntries(parts...)); err != nil {
-				return err
-			}
-			if err := sw.Close(); err != nil {
-				return err
-			}
-			sgz, err := fileSectionReader(tmpFile)
-			if err != nil {
-				return err
-			}
-			r, _, err := newVerifiableStargz(sgz)
-			if err != nil {
-				return err
-			}
 			esgzFile, err := layerFiles.TempFile("", "esgzdata")
 			if err != nil {
 				return err
 			}
-			if _, err := io.Copy(esgzFile, r); err != nil {
+			sw := NewWriter(esgzFile)
+			sw.ChunkSize = opts.chunkSize
+			if err := sw.AppendTar(readerFromEntries(parts...)); err != nil {
 				return err
 			}
-			sr, err := fileSectionReader(esgzFile)
-			if err != nil {
-				return err
-			}
-			sgzPartsMu.Lock()
-			sgzParts[i] = sr
-			sgzPartsMu.Unlock()
+			mu.Lock()
+			writers[i] = sw
+			payloads[i] = esgzFile
+			mu.Unlock()
 			return nil
 		})
 	}
@@ -124,37 +106,104 @@ func Build(tarBlob *io.SectionReader, prioritized []string, opt ...Option) (_ io
 		rErr = err
 		return nil, "", err
 	}
-	wholeBlob, tocDgst, err := combineBlobs(sgzParts...)
+	tocAndFooter, tocDgst, err := closeWithCombine(writers...)
 	if err != nil {
 		rErr = err
 		return nil, "", err
 	}
+	var rs []io.Reader
+	for _, p := range payloads {
+		fs, err := fileSectionReader(p)
+		if err != nil {
+			return nil, "", err
+		}
+		rs = append(rs, fs)
+	}
 	return readCloser{
-		Reader:    wholeBlob,
+		Reader:    io.MultiReader(append(rs, tocAndFooter)...),
 		closeFunc: layerFiles.CleanupAll,
 	}, tocDgst, nil
 }
 
-type readCloser struct {
-	io.Reader
-	closeFunc func() error
-}
-
-func (rc readCloser) Close() error {
-	return rc.closeFunc()
-}
-
-func fileSectionReader(file *os.File) (*io.SectionReader, error) {
-	info, err := file.Stat()
-	if err != nil {
-		return nil, err
+// closeWithCombine takes unclosed Writers and close them. This also returns the
+// toc that combined all Writers into.
+// Writers doesn't write TOC and footer to the underlying writers so they can be
+// combined into a single eStargz and tocAndFooter returned by this function can
+// be appended at the tail of that combined blob.
+func closeWithCombine(ws ...*Writer) (tocAndFooter io.Reader, tocDgst digest.Digest, err error) {
+	if len(ws) == 0 {
+		return nil, "", fmt.Errorf("at least one writer must be passed")
 	}
-	return io.NewSectionReader(file, 0, info.Size()), nil
+	for _, w := range ws {
+		if w.closed {
+			return nil, "", fmt.Errorf("writer must be unclosed")
+		}
+		defer func(w *Writer) { w.closed = true }(w)
+		if err := w.closeGz(); err != nil {
+			return nil, "", err
+		}
+		if err := w.bw.Flush(); err != nil {
+			return nil, "", err
+		}
+	}
+	var (
+		mtoc          = new(jtoc)
+		currentOffset int64
+	)
+	mtoc.Version = ws[0].toc.Version
+	for _, w := range ws {
+		for _, e := range w.toc.Entries {
+			// Recalculate Offset of non-empty files/chunks
+			if (e.Type == "reg" && e.Size > 0) || e.Type == "chunk" {
+				e.Offset += currentOffset
+			}
+			mtoc.Entries = append(mtoc.Entries, e)
+		}
+		if w.toc.Version > mtoc.Version {
+			mtoc.Version = w.toc.Version
+		}
+		currentOffset += w.cw.n
+	}
+
+	tocJSON, err := json.MarshalIndent(mtoc, "", "\t")
+	if err != nil {
+		return nil, "", err
+	}
+	pr, pw := io.Pipe()
+	go func() {
+		zw, _ := gzip.NewWriterLevel(pw, gzip.BestCompression)
+		tw := tar.NewWriter(zw)
+		if err := tw.WriteHeader(&tar.Header{
+			Typeflag: tar.TypeReg,
+			Name:     TOCTarName,
+			Size:     int64(len(tocJSON)),
+		}); err != nil {
+			pw.CloseWithError(err)
+			return
+		}
+		if _, err := tw.Write(tocJSON); err != nil {
+			pw.CloseWithError(err)
+			return
+		}
+		if err := tw.Close(); err != nil {
+			pw.CloseWithError(err)
+			return
+		}
+		if err := zw.Close(); err != nil {
+			pw.CloseWithError(err)
+			return
+		}
+		pw.Close()
+	}()
+	return io.MultiReader(
+		pr,
+		bytes.NewReader(footerBytes(currentOffset)),
+	), digest.FromBytes(tocJSON), nil
 }
 
 // divideEntries divides passed entries to the parts at least the number specified by the
 // argument.
-func divideEntries(entries []*tarEntry, minPartsNum int) (set [][]*tarEntry) {
+func divideEntries(entries []*entry, minPartsNum int) (set [][]*entry) {
 	var estimatedSize int64
 	for _, e := range entries {
 		estimatedSize += e.header.Size
@@ -164,22 +213,22 @@ func divideEntries(entries []*tarEntry, minPartsNum int) (set [][]*tarEntry) {
 		nextEnd = unitSize
 		offset  int64
 	)
-	set = append(set, []*tarEntry{})
+	set = append(set, []*entry{})
 	for _, e := range entries {
 		set[len(set)-1] = append(set[len(set)-1], e)
 		offset += e.header.Size
 		if offset > nextEnd {
-			set = append(set, []*tarEntry{})
+			set = append(set, []*entry{})
 			nextEnd += unitSize
 		}
 	}
 	return
 }
 
-// sort reads the specified tar blob and returns a list of tar entries.
+// sortEntries reads the specified tar blob and returns a list of tar entries.
 // If some of prioritized files are specified, the list starts from these
 // files with keeping the order specified by the argument.
-func sort(in io.ReaderAt, prioritized []string) ([]*tarEntry, error) {
+func sortEntries(in io.ReaderAt, prioritized []string) ([]*entry, error) {
 
 	// Import tar file.
 	intar, err := importTar(in)
@@ -193,7 +242,7 @@ func sort(in io.ReaderAt, prioritized []string) ([]*tarEntry, error) {
 		moveRec(l, intar, sorted)
 	}
 	if len(prioritized) == 0 {
-		sorted.add(&tarEntry{
+		sorted.add(&entry{
 			header: &tar.Header{
 				Name:     NoPrefetchLandmark,
 				Typeflag: tar.TypeReg,
@@ -202,7 +251,7 @@ func sort(in io.ReaderAt, prioritized []string) ([]*tarEntry, error) {
 			payload: bytes.NewReader([]byte{landmarkContents}),
 		})
 	} else {
-		sorted.add(&tarEntry{
+		sorted.add(&entry{
 			header: &tar.Header{
 				Name:     PrefetchLandmark,
 				Typeflag: tar.TypeReg,
@@ -218,7 +267,7 @@ func sort(in io.ReaderAt, prioritized []string) ([]*tarEntry, error) {
 
 // readerFromEntries returns a reader of tar archive that contains entries passed
 // through the arguments.
-func readerFromEntries(entries ...*tarEntry) io.Reader {
+func readerFromEntries(entries ...*entry) io.Reader {
 	pr, pw := io.Pipe()
 	go func() {
 		tw := tar.NewWriter(pw)
@@ -240,7 +289,7 @@ func readerFromEntries(entries ...*tarEntry) io.Reader {
 
 func importTar(in io.ReaderAt) (*tarFile, error) {
 	tf := &tarFile{}
-	pw, err := positionwatcher.NewPositionWatcher(in)
+	pw, err := newCountReader(in)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to make position watcher")
 	}
@@ -266,9 +315,9 @@ func importTar(in io.ReaderAt) (*tarFile, error) {
 		if _, ok := tf.get(h.Name); ok {
 			return nil, fmt.Errorf("Duplicated entry(%q) is not supported", h.Name)
 		}
-		tf.add(&tarEntry{
+		tf.add(&entry{
 			header:  h,
-			payload: io.NewSectionReader(in, pw.CurrentPos(), h.Size),
+			payload: io.NewSectionReader(in, pw.currentPos(), h.Size),
 		})
 	}
 
@@ -290,19 +339,19 @@ func moveRec(name string, in *tarFile, out *tarFile) {
 	}
 }
 
-type tarEntry struct {
+type entry struct {
 	header  *tar.Header
 	payload io.ReadSeeker
 }
 
 type tarFile struct {
-	index  map[string]*tarEntry
-	stream []*tarEntry
+	index  map[string]*entry
+	stream []*entry
 }
 
-func (f *tarFile) add(e *tarEntry) {
+func (f *tarFile) add(e *entry) {
 	if f.index == nil {
-		f.index = make(map[string]*tarEntry)
+		f.index = make(map[string]*entry)
 	}
 	f.index[e.header.Name] = e
 	f.stream = append(f.stream, e)
@@ -312,7 +361,7 @@ func (f *tarFile) remove(name string) {
 	if f.index != nil {
 		delete(f.index, name)
 	}
-	var filtered []*tarEntry
+	var filtered []*entry
 	for _, e := range f.stream {
 		if e.header.Name == name {
 			continue
@@ -322,7 +371,7 @@ func (f *tarFile) remove(name string) {
 	f.stream = filtered
 }
 
-func (f *tarFile) get(name string) (e *tarEntry, ok bool) {
+func (f *tarFile) get(name string) (e *entry, ok bool) {
 	if f.index == nil {
 		return nil, false
 	}
@@ -330,6 +379,109 @@ func (f *tarFile) get(name string) (e *tarEntry, ok bool) {
 	return
 }
 
-func (f *tarFile) dump() []*tarEntry {
+func (f *tarFile) dump() []*entry {
 	return f.stream
+}
+
+type readCloser struct {
+	io.Reader
+	closeFunc func() error
+}
+
+func (rc readCloser) Close() error {
+	return rc.closeFunc()
+}
+
+func fileSectionReader(file *os.File) (*io.SectionReader, error) {
+	info, err := file.Stat()
+	if err != nil {
+		return nil, err
+	}
+	return io.NewSectionReader(file, 0, info.Size()), nil
+}
+
+func newTempFiles() *tempFiles {
+	return &tempFiles{}
+}
+
+type tempFiles struct {
+	files   []*os.File
+	filesMu sync.Mutex
+}
+
+func (tf *tempFiles) TempFile(dir, pattern string) (*os.File, error) {
+	f, err := ioutil.TempFile(dir, pattern)
+	if err != nil {
+		return nil, err
+	}
+	tf.filesMu.Lock()
+	tf.files = append(tf.files, f)
+	tf.filesMu.Unlock()
+	return f, nil
+}
+
+func (tf *tempFiles) CleanupAll() (allErr error) {
+	tf.filesMu.Lock()
+	defer tf.filesMu.Unlock()
+	for _, f := range tf.files {
+		if err := f.Close(); err != nil {
+			allErr = multierror.Append(allErr, err)
+		}
+		if err := os.Remove(f.Name()); err != nil {
+			allErr = multierror.Append(allErr, err)
+		}
+	}
+	tf.files = nil
+	return nil
+}
+
+func newCountReader(r io.ReaderAt) (*countReader, error) {
+	pos := int64(0)
+	return &countReader{r: r, cPos: &pos}, nil
+}
+
+type countReader struct {
+	r    io.ReaderAt
+	cPos *int64
+
+	mu sync.Mutex
+}
+
+func (cr *countReader) Read(p []byte) (int, error) {
+	cr.mu.Lock()
+	defer cr.mu.Unlock()
+
+	n, err := cr.r.ReadAt(p, *cr.cPos)
+	if err == nil {
+		*cr.cPos += int64(n)
+	}
+	return n, err
+}
+
+func (cr *countReader) Seek(offset int64, whence int) (int64, error) {
+	cr.mu.Lock()
+	defer cr.mu.Unlock()
+
+	switch whence {
+	default:
+		return 0, fmt.Errorf("Unknown whence: %v", whence)
+	case io.SeekStart:
+	case io.SeekCurrent:
+		offset += *cr.cPos
+	case io.SeekEnd:
+		return 0, fmt.Errorf("Unsupported whence: %v", whence)
+	}
+
+	if offset < 0 {
+		return 0, fmt.Errorf("invalid offset")
+	}
+	*cr.cPos = offset
+	return offset, nil
+}
+
+func (cr *countReader) currentPos() int64 {
+	cr.mu.Lock()
+	defer cr.mu.Unlock()
+
+	return *cr.cPos
 }
